requirements.txt
# Core Bot Framework
python-telegram-bot[job-queue]==21.6

# Database
SQLAlchemy>=2.0.0,<3.0.0
psycopg2-binary>=2.9.9  # PostgreSQL driver

# Environment & Configuration
python-dotenv>=1.0.0

# Timezone handling
pytz>=2024.1

# HTTP (required by telegram bot)
httpx>=0.27.0

# Logging
python-json-logger>=2.0.0
